TensileCompanion/requirements.txt
pyserial>=3.5
matplotlib>=3.5.0
reportlab>=4.0.0
